parsing_documents/parsePDFs02/requirements.txt
pip
pandas
numpy
pathlib
tabula-py